mcp-chatwoot/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
# pymongo>=4.3.3  # Comentado para remover dependência
# redis>=4.5.4    # Comentado para remover dependência
httpx>=0.26.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
fastmcp>=0.1.0
cryptography>=40.0.2
